Fix most warnings in MusicBrainz.
