+++ MediaBrowser.Providers/Plugins/MusicBrainz/Plugin.cs
@@ -11,6 +11,16 @@ namespace MediaBrowser.Providers.Plugins.MusicBrainz
 {
     public class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
     {
+        public const string DefaultServer = "https://musicbrainz.org";
+
+        public const long DefaultRateLimit = 2000u;
+
+        public Plugin(IApplicationPaths applicationPaths, IXmlSerializer xmlSerializer)
+            : base(applicationPaths, xmlSerializer)
+        {
+            Instance = this;
+        }
+
         public static Plugin Instance { get; private set; }
 
         public override Guid Id => new Guid("8c95c4d2-e50c-4fb0-a4f3-6c06ff0f9a1a");
@@ -19,18 +29,8 @@ namespace MediaBrowser.Providers.Plugins.MusicBrainz
 
         public override string Description => "Get artist and album metadata from any MusicBrainz server.";
 
-        public const string DefaultServer = "https://musicbrainz.org";
-
-        public const long DefaultRateLimit = 2000u;
-
         // TODO remove when plugin removed from server.
         public override string ConfigurationFileName => "Jellyfin.Plugin.MusicBrainz.xml";
-
-        public Plugin(IApplicationPaths applicationPaths, IXmlSerializer xmlSerializer)
-            : base(applicationPaths, xmlSerializer)
-        {
-            Instance = this;
-        }
 
         public IEnumerable<PluginPageInfo> GetPages()
         {
